# ScFrameWork/docs/openapi-starter.json
{
  "components": {
    "schemas": {
...
      "EmptyCommand": {},
...
      "OperationMessageItem": {
        "properties": {
          "completedAt": {
            "format": "date-time",
            "type": ["string", "null"]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "dispatchAttempts": {
            "format": "int32",
            "maximum": 5,
            "minimum": 0,
            "type": "integer"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "lastFailureCode": {
            "maxLength": 32,
            "type": ["string", "null"]
          },
          "nextAttemptAt": {
            "format": "date-time",
            "type": "string"
          },
          "publishedAt": {
            "format": "date-time",
            "type": ["string", "null"]
          },
          "state": {
            "enum": ["PENDING", "CLAIMED", "PUBLISHED", "COMPLETED", "DEAD"],
            "type": "string"
          },
          "type": {
            "enum": ["SECURITY_AUDIT", "FILE_DELETE", "MESSAGE_DEMO"],
            "type": "string"
          }
        },
        "required": [
          "completedAt",
          "createdAt",
          "dispatchAttempts",
          "eventId",
          "lastFailureCode",
          "nextAttemptAt",
          "publishedAt",
          "state",
          "type"
        ],
        "type": "object"
      },
      "OperationMessagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationMessageItem"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "maximum": 100,
            "minimum": 1,
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["items", "page", "size", "total"],
        "type": "object"
      },
...
      "StoredBlob": {
        "properties": {
          "key": {
            "type": "string"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "session": {
        "description": "같은 브라우저의 로그인 세션. HttpOnly 쿠키를 Authorize 입력으로 설정하지 않는다.",
        "in": "cookie",
        "name": "JSESSIONID",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "sc-starter-app",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operations/messages": {
      "get": {
        "operationId": "scOperations_get_messages",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 1000000,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "enum": ["PENDING", "CLAIMED", "PUBLISHED", "COMPLETED", "DEAD"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OperationMessagePage"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["starter-message-controller"]
      }
    },
    "/api/operations/messages/demo": {
      "post": {
        "operationId": "scOperations_post_messages_demo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommand"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OperationMessageItem"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "tags": ["starter-message-controller"]
      }
    },
    "/api/operations/messages/{eventId}/retry": {
      "post": {
        "operationId": "scOperations_post_messages_eventId_retry",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyCommand"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OperationMessageItem"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "tags": ["starter-message-controller"]
      }
    },
    "/api/storage-demo": {
      "post": {
        "operationId": "upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StoredBlob"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["storage-demo-controller"]
      }
    },
    "/api/storage-demo/{key}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": ["storage-demo-controller"]
      },
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": ["storage-demo-controller"]
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ],
  "servers": [
    {
      "description": "같은 출처의 실행 서버",
      "url": "/"
    }
  ]
}